# Avoid allocation of dynamic config constraint precedence