--- a/common/dynamicconfig/setting_gen.go
+++ b/common/dynamicconfig/setting_gen.go
@@ -778,13 +778,14 @@ type TypedPropertyFn[T any] func() T
 
 func (s GlobalTypedSetting[T]) Get(c *Collection) TypedPropertyFn[T] {
 	return func() T {
+		prec := []Constraints{{}}
 		return matchAndConvert(
 			c,
 			s.key,
 			s.def,
 			s.cdef,
 			s.convert,
-			precedenceGlobal(),
+			prec,
 		)
 	}
 }
@@ -852,13 +853,14 @@ type TypedPropertyFnWithNamespaceFilter[T any] func(namespace string) T
 
 func (s NamespaceTypedSetting[T]) Get(c *Collection) TypedPropertyFnWithNamespaceFilter[T] {
 	return func(namespace string) T {
+		prec := []Constraints{{Namespace: namespace}, {}}
 		return matchAndConvert(
 			c,
 			s.key,
 			s.def,
 			s.cdef,
 			s.convert,
-			precedenceNamespace(namespace),
+			prec,
 		)
 	}
 }
@@ -926,13 +928,14 @@ type TypedPropertyFnWithNamespaceIDFilter[T any] func(namespaceID string) T
 
 func (s NamespaceIDTypedSetting[T]) Get(c *Collection) TypedPropertyFnWithNamespaceIDFilter[T] {
 	return func(namespaceID string) T {
+		prec := []Constraints{{NamespaceID: namespaceID}, {}}
 		return matchAndConvert(
 			c,
 			s.key,
 			s.def,
 			s.cdef,
 			s.convert,
-			precedenceNamespaceID(namespaceID),
+			prec,
 		)
 	}
 }
@@ -1000,13 +1003,20 @@ type TypedPropertyFnWithTaskQueueFilter[T any] func(namespace string, taskQueue 
 
 func (s TaskQueueTypedSetting[T]) Get(c *Collection) TypedPropertyFnWithTaskQueueFilter[T] {
 	return func(namespace string, taskQueue string, taskQueueType enumspb.TaskQueueType) T {
+		prec := []Constraints{
+			{Namespace: namespace, TaskQueueName: taskQueue, TaskQueueType: taskQueueType},
+			{Namespace: namespace, TaskQueueName: taskQueue},
+			{TaskQueueName: taskQueue},
+			{Namespace: namespace},
+			{},
+		}
 		return matchAndConvert(
 			c,
 			s.key,
 			s.def,
 			s.cdef,
 			s.convert,
-			precedenceTaskQueue(namespace, taskQueue, taskQueueType),
+			prec,
 		)
 	}
 }
@@ -1074,13 +1084,14 @@ type TypedPropertyFnWithShardIDFilter[T any] func(shardID int32) T
 
 func (s ShardIDTypedSetting[T]) Get(c *Collection) TypedPropertyFnWithShardIDFilter[T] {
 	return func(shardID int32) T {
+		prec := []Constraints{{ShardID: shardID}, {}}
 		return matchAndConvert(
 			c,
 			s.key,
 			s.def,
 			s.cdef,
 			s.convert,
-			precedenceShardID(shardID),
+			prec,
 		)
 	}
 }
@@ -1148,13 +1159,14 @@ type TypedPropertyFnWithTaskTypeFilter[T any] func(taskType enumsspb.TaskType) T
 
 func (s TaskTypeTypedSetting[T]) Get(c *Collection) TypedPropertyFnWithTaskTypeFilter[T] {
 	return func(taskType enumsspb.TaskType) T {
+		prec := []Constraints{{TaskType: taskType}, {}}
 		return matchAndConvert(
 			c,
 			s.key,
 			s.def,
 			s.cdef,
 			s.convert,
-			precedenceTaskType(taskType),
+			prec,
 		)
 	}
 }
@@ -1222,13 +1234,19 @@ type TypedPropertyFnWithDestinationFilter[T any] func(namespace string, destinat
 
 func (s DestinationTypedSetting[T]) Get(c *Collection) TypedPropertyFnWithDestinationFilter[T] {
 	return func(namespace string, destination string) T {
+		prec := []Constraints{
+			{Namespace: namespace, Destination: destination},
+			{Destination: destination},
+			{Namespace: namespace},
+			{},
+		}
 		return matchAndConvert(
 			c,
 			s.key,
 			s.def,
 			s.cdef,
 			s.convert,
-			precedenceDestination(namespace, destination),
+			prec,
 		)
 	}
 }