## What changed?
Inline slices representation constraint precedence in generated code.

## Why?
This saves an allocation per dynamic config call.

## How did you test it?
Existing tests + bench.