@dnr 2024-07-08
benchstat (just on my laptop, timing may not be representative)
```
goos: linux
goarch: amd64
pkg: go.temporal.io/server/common/dynamicconfig
cpu: Intel(R) Xeon(R) W-10855M CPU @ 2.80GHz
                                 │  before1.txt  │              after1.txt              │
                                 │    sec/op     │    sec/op     vs base                │
Collection/global_int-12           259.40n ± 23%   42.20n ±  6%  -83.73% (p=0.000 n=10)
Collection/namespace_int-12         382.2n ± 12%   130.9n ±  8%  -65.76% (p=0.000 n=10)
Collection/taskqueue_int-12         320.4n ± 77%   157.8n ± 12%  -50.76% (p=0.000 n=10)
Collection/single_default-12        411.2n ± 53%   119.1n ±  1%  -71.04% (p=0.000 n=10)
Collection/structured_default-12    724.2n ± 16%   210.2n ±  3%  -70.97% (p=0.000 n=10)
geomean                             393.7n         116.9n        -70.31%

                                 │ before1.txt  │               after1.txt                │
                                 │     B/op     │    B/op     vs base                     │
Collection/global_int-12             80.00 ± 0%    0.00 ± 0%  -100.00% (p=0.000 n=10)
Collection/namespace_int-12        164.000 ± 0%   4.000 ± 0%   -97.56% (p=0.000 n=10)
Collection/taskqueue_int-12        420.000 ± 0%   4.000 ± 0%   -99.05% (p=0.000 n=10)
Collection/single_default-12         416.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
Collection/structured_default-12     416.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
geomean                              248.8                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                                 │ before1.txt │               after1.txt                │
                                 │  allocs/op  │ allocs/op   vs base                     │
Collection/global_int-12            1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
Collection/namespace_int-12         1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
Collection/taskqueue_int-12         1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
Collection/single_default-12        1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
Collection/structured_default-12    1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean                             1.000                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```